Updated auto-responder to allow duplicate responses, handle deleting.

+++ cogs/autoresponder/autoresponder.py
@@ -1,5 +1,5 @@
-import random
 import requests
+from random import choice
 from pyquery import PyQuery as pq
 from discord.ext import commands
 from tabulate import tabulate
@@ -24,7 +24,8 @@ class AutoResponder(commands.Cog):
         """
         await db.open()
         try:
-            response = await Response.get(shortcut=args[0])
+            responses = await Response.filter(shortcut=args[0])
+            response = choice(responses)
             await context.send(response.text)
         
         except DoesNotExist:
@@ -36,14 +37,14 @@ class AutoResponder(commands.Cog):
     @commands.has_any_role(*ROLES_CAN_EDIT)
     async def add(self, context, *args):
         """
-        Add an autoresponder: show add popranked "It's time to pop ranked!"
+        [MODS ONLY] Add response: show add popranked "It's time to pop ranked!"
         """
         await db.open()
         shortcut = args[0]
         text = ' '.join(args[1:])
         response = Response(shortcut=shortcut, text=text)
         await response.save()
-        await context.send(f'Added response `!{shortcut} {text}`')
+        await context.send(f'Added response `{shortcut} {text}`')
         await db.close()
         
 
@@ -51,16 +52,42 @@ class AutoResponder(commands.Cog):
     @commands.has_any_role(*ROLES_CAN_EDIT)
     async def delete(self, context, *args):
         """
-        Delete an autoresponder: show delete popranked
+        [MODS ONLY] Delete an autoresponder: show delete popranked
         """
         await db.open()
         shortcut = args[0]
 
         try:
-            response = await Response.get(shortcut=shortcut)
-            await response.delete()                
-            await context.send(f'`{shortcut}` **deleted.**')
-        
+            responses = await Response.filter(shortcut=shortcut).order_by('id')
+
+            if len(responses) == 1:
+                await responses.delete()
+                await context.send(f'`{shortcut}` **deleted.**')
+            elif len(responses) > 1:
+                if len(args) == 1:
+                    headers = ['Index', 'Shortcut', 'Response']
+                    table = []
+                    for idx, response in enumerate(responses):
+                        text = response.text
+                        text = text.replace ('\n', ' ')
+
+                        if len(text) > 50:
+                            text = text[:49] + '...'
+                        table.append([idx, response.shortcut, text]) 
+
+                    table_data = tabulate(table, headers=headers, tablefmt='presto')
+                    msg = f'__Here is a list of all auto-responders with that shorcut__:\n```\n{table_data}\n```'                
+                    msg += f'Enter `show delete {shortcut} <index>` to delete a response.'
+                    await context.send(msg)
+                if len(args) == 2:
+                    try:
+                        index = int(args[1])
+                        response = responses[index]
+                        await response.delete()
+                        await context.send(f'`{shortcut}[{index}]` **deleted.**')
+                    except (IndexError, ValueError):
+                        pass
+
         except DoesNotExist:
             pass
 
@@ -69,22 +96,35 @@ class AutoResponder(commands.Cog):
     @show.command()
     async def list(self, context, *args):
         """
-        List all auto responders available for use.
+        List all auto responders available for use. 'list mobile' and 'list full' work too.
         """
         await db.open()
         responses = await Response.all()
         msg = '__Here all autoresponders__'
         headers = ['Shortcut', 'Response']
+        tablefmt = 'presto'
         table = []
+
+        # Accomodate different screen sizes on request.
+        width = 50
+
+        if args:
+            if args[0] == 'mobile':
+                width = 15
+                tablefmt = 'simple'
+            elif args[0] == 'full':
+                width = 1024
+                tablefmt = 'simple'
+
         for response in responses:
             text = response.text
             text = text.replace ('\n', ' ')
 
-            if len(text) > 50:
-                text = text[:49] + '...'
+            if len(text) > width:
+                text = text[:width-1] + '...'
             table.append([response.shortcut, text]) 
 
-        table_data = tabulate(table, headers=headers, tablefmt='presto')
+        table_data = tabulate(table, headers=headers, tablefmt=tablefmt)
         msg = f'__Here is a list of all auto-responders__:\n```\n{table_data}\n```'
         await context.send(msg)
         await db.close()
